Changed the SIMULATED_TIME timing to real time analog, getting rid of explicit integer time windows. Changed the defaults of CYST and META actions, and routing to sensible millisecond values.

+++ cyst/platform/network/node.py
@@ -74,7 +74,7 @@ class NodeImpl(Node):
         i.set_index(index)
         return index
 
-    async def process_message(self, message, delay: int) -> int:
+    async def process_message(self, message, delay: float) -> int:
         if message.type == MessageType.ACK:
             return 0
